Export SHP
ShapeFileWriter now takes a DataResultTable and export the data in
a single ZIP bytearray with one shapefile for exising geometry type.

I've tested to load the zip with QGIS and it opens directly all the shapefiles included in the zip.

After having experimented with org.arakhne.afc.advanced shapefile
library I went back and using the more used and standard library
geotools.
In that context geotools is non optimal because it works using
Java Files, so hard to make it transactional, arakhne is supporting
bytestream, but is not very documented and lacks on geographic
datatype (like Polygons with holes).
A first attempt to use geotools without writing on the filesystem
was done with https://google.github.io/jimfs/ but geotools in
enforcing local files so no luck.

Then in order to use geotools I had to solve some conflict in the JTS
library, ensuring that slick-pg geotrellis and geotools are using
the same version. I did that using the great sbt dependencyBrowseTree

Fixes: #180